backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
click==8.3.2
distro==1.9.0
fastapi==0.135.3
filelock==3.25.2
fsspec==2026.3.0
grpcio==1.80.0
h11==0.16.0
h2==4.3.0
hf-xet==1.4.3
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.10.1
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jiter==0.14.0
joblib==1.5.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
openai==2.32.0
packaging==26.0
portalocker==3.2.0
protobuf==7.34.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
PyMuPDF==1.27.2.2
python-dotenv==1.2.2
python-multipart==0.0.26
PyYAML==6.0.3
qdrant-client==1.17.1
regex==2026.4.4
rich==15.0.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.4.0
shellingham==1.5.4
sniffio==1.3.1
starlette==1.0.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==5.5.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
